// src/interfaces/userStatus.ts
export enum QlsUserStatus {
    Weakpassword = "WEAKPASSWORD",
    EmailInUse = "EMAILINUSE", 
    InvalidEmail = "INVALIDEMAIL",
    OperationNotAllowed = "OPERATIONNOTALLOWED",
    UserDisabled = "USERDISABLED",
    UserNotFound = "USERNOTFOUND",
    UserNotCreated = "USERNOTCREATED",
    WrongPassword = "WRONGPASSWORD",
    Ok = "OK",
    Error = "ERROR",
    LoggedIn = "LOGGEDIN",
    LoggedOut = "LOGGEDOUT"
}